+++ parser/src/lib.rs
@@ -41,4 +41,4 @@ mod scanner;
 
 pub use crate::input::{str::StrInput, BufferedInput, Input};
 pub use crate::parser::{Event, EventReceiver, Parser, SpannedEventReceiver, Tag};
-pub use crate::scanner::{Marker, ScanError, Span, TScalarStyle};
+pub use crate::scanner::{Marker, ScalarStyle, ScanError, Span};