+++ saphyr/src/annotated.rs
@@ -27,7 +27,7 @@ use std::{
 };
 
 use hashlink::LinkedHashMap;
-use saphyr_parser::{TScalarStyle, Tag};
+use saphyr_parser::{ScalarStyle, Tag};
 
 use crate::Scalar;
 
@@ -68,7 +68,7 @@ where
     HashKey: Eq + std::hash::Hash + std::borrow::Borrow<Node> + From<Node>,
 {
     /// The raw string from the input.
-    Representation(Cow<'input, str>, TScalarStyle, Option<Tag>),
+    Representation(Cow<'input, str>, ScalarStyle, Option<Tag>),
     /// The resolved value from the representation.
     Value(Scalar<'input>),
     /// YAML sequence, can be accessed as a `Vec`.